// src/app/models/register-user-request.model.ts
export interface RegisterUserRequest {
  firstName: string;
  middleName: string;
  lastName: string;
  gender: string;
  dateOfBirth: string;
  age: number;
  address: string;
  city: string;
  pinCode: string;
  mobileNumber: string;
  email: string;
  password: string;
  whatsappNumber: string;
  emergencyNumber: string;
  idProofType: string;
  idProofNumber: string;
  idProofPath: string;
  memberRelation: string;
  daysAttending: string;
  accommodationRequired: string;
  mealsRequired: string;
  healthProblem: string;
  remarks: string;
  noOfMembers: number;
  isActive: boolean;
  isLocked: boolean;
  createdBy: string;
  createdDateTime: string;
  updatedBy: string;
  updatedDatetime: string;
  corpoName: string;
  branch: string;
  landlineNumber: string;
  userType: number;
  eventId: number;
  isSelected: boolean;
}